+++ contracts/zap/safeSwapBNB.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.4;
 
-import "@pancakeswap/pancake-swap-lib/contracts/token/BEP20/BEP20.sol";
-import "@pancakeswap/pancake-swap-lib/contracts/token/BEP20/SafeBEP20.sol";
+import "../library/bep20/BEP20.sol";
+import "../library/bep20/SafeBEP20.sol";
 import "@pancakeswap/pancake-swap-lib/contracts/math/SafeMath.sol";
 
 import "../library/SafeToken.sol";
@@ -16,7 +16,6 @@ contract safeSwapBNB {
 
     address private constant WBNB = 0xbb4CdB9CBd36B01bD1cBaEBF2De08d9173bc095c;
 
-
     /* ========== CONSTRUCTOR ========== */
 
     constructor() public {}
@@ -25,14 +24,16 @@ contract safeSwapBNB {
 
     /* ========== FUNCTIONS ========== */
 
-    function withdraw(uint amount) external {
-        require(IBEP20(WBNB).balanceOf(msg.sender) >= amount, "Not enough Tokens!");
+    function withdraw(uint256 amount) external {
+        require(
+            IBEP20(WBNB).balanceOf(msg.sender) >= amount,
+            "Not enough Tokens!"
+        );
 
         IBEP20(WBNB).transferFrom(msg.sender, address(this), amount);
 
         IWETH(WBNB).withdraw(amount);
 
         SafeToken.safeTransferETH(msg.sender, amount);
-
     }
 }
